backend/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2
langchain==0.1.9
langchain-anthropic==0.1.4
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1

# NLP Processing
spacy==3.7.2
nltk==3.8.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis (for caching and task queue)
redis==5.0.1
celery==5.3.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Environment Variables
python-dotenv==1.0.1

# Validation
email-validator==2.1.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2
pillow==10.2.0

# Audio Transcription
whisper==1.1.10
pydub==0.25.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
phonenumbers==8.13.29

# Testing
pytest==8.0.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Security
cryptography==42.0.2
bcrypt==4.1.2

# Monitoring and Logging
sentry-sdk==1.40.0
python-json-logger==2.0.7

# Rate Limiting
Flask-Limiter==3.5.0

# Background Jobs
APScheduler==3.10.4

# Communication
twilio==8.13.0
imapclient==2.3.1
python-multipart==0.0.9
